move discovery schemas to platforms
