+++ homeassistant/components/matter/switch.py
@@ -4,8 +4,13 @@ from __future__ import annotations
 from typing import Any
 
 from chip.clusters import Objects as clusters
+from matter_server.client.models import device_types
 
-from homeassistant.components.switch import SwitchEntity
+from homeassistant.components.switch import (
+    SwitchDeviceClass,
+    SwitchEntity,
+    SwitchEntityDescription,
+)
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import Platform
 from homeassistant.core import HomeAssistant, callback
@@ -13,6 +18,7 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .entity import MatterEntity
 from .helpers import get_matter
+from .models import MatterDiscoverySchema
 
 
 async def async_setup_entry(
@@ -50,3 +56,19 @@ class MatterSwitch(MatterEntity, SwitchEntity):
         self._attr_is_on = self.get_matter_attribute_value(
             self._entity_info.primary_attribute
         )
+
+
+# Discovery schema(s) to map Matter Attributes to HA entities
+DISCOVERY_SCHEMAS = [
+    MatterDiscoverySchema(
+        platform=Platform.SWITCH,
+        entity_description=SwitchEntityDescription(
+            key="MatterPlug", device_class=SwitchDeviceClass.OUTLET
+        ),
+        entity_class=MatterSwitch,
+        required_attributes=(clusters.OnOff.Attributes.OnOff,),
+        # restrict device type to prevent discovery by light
+        # platform which also uses OnOff cluster
+        device_type=(device_types.OnOffPlugInUnit,),
+    ),
+]
